rename Word to Substr and add trait impls
* Rename the increasingly misnamed Word to Substr, reflecting its use to
  represent a sequence of words within a larger phrase.
* impl PartialEq<str> for Substr
* impl Deref<Target = str> for Substr
